Java code activity to write to a file.
This java code can be used in a TIBCO Java code activity, and output the "error" using an error logger from log4j configuration. Error logger is a rolling file that writes to a specified file. This can be used together with TibcoKafkaConnector code since this can be the error handling, in case we need the scenario that Kafka fails, to write the messages locally somewhere as a fallback.